pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visual-memory"
version = "0.1.0"
description = "Visual memory system with object detection for FindIt app"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "pillow-heif>=0.10.0",
]

[project.scripts]
remember-mode = "remember_mode.main:main"
scan-mode = "scan_mode.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["remember_mode*", "scan_mode*", "utils*"]
